- "I don't like to give away pizzas for nothing ;-)"-performance patches   for mapTexture()
  Results in almost 20% better overall performance.

- Better testcase for MarbleTest

- Adding airports ( not done yet ) 

- Small patches by Benoit Sigoure 
  (which result in maybe 1% improved performance here)

- initial changes to tinywebbrowser to get image loading
  for Wikipedia support get done in the future.


svn path=/trunk/KDE/kdeedu/marble/; revision=685586

+++ src/lib/httpfetchfile.h
@@ -17,6 +17,7 @@
 #ifndef HTTPFETCHFILE_H
 #define HTTPFETCHFILE_H
 
+#include <QtGlobal>
 #include <QtCore/QFile>
 #include <QtCore/QObject>
 #include <QtCore/QUrl>
@@ -43,7 +44,7 @@ struct HttpJob
     QString  targetDirString;
     QFile*   targetFile;
 
-    int      initiatorId;
+    qint64   initiatorId;
     Priority priority;
     Status   status;
 };
